rt: move task::Id into its own file
This is a minor internal cleanup.
